Move 'SequenceInformation' class into dataclass file
